requirements.txt
# Core Trading APIs
schwab-py>=0.1.0
yfinance>=0.2.0

# Database
sqlalchemy>=1.4.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web Framework (Dashboard)
flask>=2.3.0
flask-limiter>=3.5.0
jinja2>=3.1.0

# API Communication
requests>=2.31.0
tenacity>=8.2.0
python-dotenv>=1.0.0

# Telegram Notifications
python-telegram-bot>=20.0

# Google Sheets Integration
gspread>=5.10.0

# Date/Time
pytz>=2023.3

# Technical Analysis (Optional)
scipy>=1.10.0
trendln>=0.7.0  # For support/resistance level calculation

# Visualization
plotly>=5.18.0
kaleido>=0.2.1  # For static image export (optional)

# Testing (Development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Optional: ML/AI (for future enhancements)
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Optional: Advanced Features
openai>=0.27.0  # For GPT integration
newsapi-python>=0.2.7  # For news sentiment
polygon-api-client>=1.0.0  # For alternative market data
ta>=0.11.0  # For additional technical indicators
reportlab>=4.0.0  # For PDF reports
streamlit>=1.25.0  # For alternative dashboard
